+++ geo/style.py
@@ -4,17 +4,20 @@ Module who contain necessary to play with style of any svg element.
 
 class Style:
 
-	def __init__(self, fill_color=None, stroke_linecaps=None, stroke_color=None, stroke_width=2, opacity=1, others_rules=None):
+	def __init__(self, fill_color=None, stroke_color=None, stroke_width=2, stroke_linecaps=None,
+	             opacity=1, others_rules=None, custom=False):
 		""" Create a style for svg
 
 		Args:
 			fill_color      (str)  : The color to fill shape. Default is black.
-			stroke_linecaps (str)  : The type of linecaps border. Default is none.
 			stroke_color    (str)  : The color of stroke of the shape. Default is blue.
-			stroke_width    (int)  : The size of stroke of the shape. Default is 2.
+			stroke_width    (int)  : The size of stroke of the shape. Default is no stroke.
+			stroke_linecaps (str)  : The type of linecaps border. Default is None.
 			opacity         (int)  : The opacity of shape. Default is 1.
 			others_rules    (list) : A list of some others rules. Default is no others rules.
+			custom          (bool) : A bool who indicate if we set a default style or if it's custom.
 		"""
+		self.custom = custom
 		# Set if we use fill, stroke, opacity... and set it
 		self.fill_color = fill_color
 		self.stroke_linecaps = stroke_linecaps
@@ -47,15 +50,16 @@ class Style:
 		"""
 		self.others_rules.clear()
 
-	def set(self, fill_color=None, stroke_linecaps=None, stroke_color=None, stroke_width=None, opacity=None):
+	def set(self, fill_color=None, stroke_color=None, stroke_width=None, stroke_linecaps=None,
+	        opacity=None, custom=None):
 		"""Set style of given values
 
 		Args:
-			fill_color      (str)  : The color to fill shape. Default is black.
-			stroke_linecaps (str)  : The type of linecaps border. Default is none.
-			stroke_color    (str)  : The color of stroke of the shape. Default is blue.
-			stroke_width    (int)  : The size of stroke of the shape. Default is 2.
-			opacity         (int)  : The opacity of shape. Default is 1.
+			fill_color   (str)    : The color to fill shape.
+			stroke_color (str)    : The color of stroke of the shape.
+			stroke_width (float)  : The size of stroke of the shape
+			opacity      (float)  : The opacity of shape.
+			custom       (bool)   : A bool who indicate if we set a default style or if it's custom.
 		"""
 		if fill_color:
 			self.fill_color = fill_color